rename safegraph folder to mapbox
