+++ examples/website/mapbox/app.js
