+++ examples/website/mapbox/index.html
