+++ examples/website/mapbox/vite.config.js
